Provide active goal stats.
This commit exposes an endpoint which serves up the list of active goals
within Hustle. This endpoint will be used by a google sheet to determine
which goals to display on the volunteer sign up sheet.

Of note: Hustle is built on top of Parse. They expose their Parse
instance directly, allowing us to query it like a database.